feat: add more convenience QueryConstraint's for handling ParseObjects
